big mess of a re-factor
