+++ api/app/tests/fba_calc/test_fba_error.py
@@ -7,7 +7,7 @@ from pytest_bdd import scenario, given, then, parsers
 from app import configure_logging
 from app.schemas.fba_calc import FuelTypeEnum
 from app.utils.time import get_hour_20_from_date
-from app.fba_calculator import calculate_fire_behaviour_advisory, FBACalculatorWeatherStation
+from app.fbp.fba_calculator import calculate_fire_behaviour_advisory, FBACalculatorWeatherStation
 from app.utils.redapp import FBPCalculateStatisticsCOM
 from app.tests.common import str2float
 from app.tests.fba_calc import (check_metric, acceptable_margin_of_error,
