+++ api/app/wildfire_one/schema_parsers.py
@@ -5,19 +5,12 @@ import logging
 from datetime import datetime, timezone
 from typing import Generator, List, Optional
 from app.db.models.observations import HourlyActual
-from app.schemas.fba_calc import FuelTypeEnum
-from app.schemas.shared import FuelType
 from app.schemas.stations import WeatherStation
-from app.utils import cffdrs
-from app.utils import c7b
 from app.utils.dewpoint import compute_dewpoint
 from app.data.ecodivision_seasons import EcodivisionSeasons
 from app.schemas.observations import WeatherReading
 from app.db.models.forecasts import NoonForecast
-from app.schemas.hfi_calc import StationDaily
-from app.utils.fuel_types import FUEL_TYPE_DEFAULTS
-from app.fba_calculator import calculate_cfb, get_fire_size, get_fire_type
-from app.utils.time import get_julian_date_now, get_utc_now
+from app.utils.time import get_utc_now
 from app.wildfire_one.util import is_station_valid, is_station_fire_zone_valid, get_zone_code_prefix
 from app.wildfire_one.validation import get_valid_flags
 from app.schemas.fba import FireCentre, FireCenterStation
@@ -151,271 +144,6 @@ def parse_hourly(hourly) -> WeatherReading:
         fwi=hourly.get('fireWeatherIndex', None),
         observation_valid=hourly.get('observationValidInd'),
         observation_valid_comment=hourly.get('observationValidComment')
-    )
-
-
-def calculate_intensity_group(hfi: float) -> int:
-    """ Returns a 1-5 integer value indicating Intensity Group based on HFI.
-    Intensity groupings are:
-
-    HFI             IG
-    0-499            1
-    500-999          2
-    1000-1999        3
-    2000-3999        4
-    4000+            5
-    """
-    if hfi < 500:
-        return 1
-    if hfi < 1000:
-        return 2
-    if hfi < 2000:
-        return 3
-    if hfi < 4000:
-        return 4
-    return 5
-
-
-class FireBehaviourPrediction:
-    """ Structure for storing fire behaviour prediction data. """
-
-    def __init__(self, ros: float,
-                 hfi: float, intensity_group,
-                 sixty_minute_fire_size: float,
-                 fire_type) -> None:
-        self.ros = ros
-        self.hfi = hfi
-        self.intensity_group = intensity_group
-        self.sixty_minute_fire_size = sixty_minute_fire_size
-        self.fire_type = fire_type
-
-
-def calculate_fire_behaviour_prediction_using_cffdrs(  # pylint: disable=too-many-arguments
-        latitude: float,
-        longitude: float,
-        elevation: float,
-        fuel_type: FuelTypeEnum,
-        bui: float,
-        ffmc: float,
-        cc: float,  # pylint: disable=invalid-name
-        pc: float,  # pylint: disable=invalid-name
-        wind_speed: float,
-        isi: float,
-        pdf: float,
-        cbh: float,
-        cfl: float):
-    """ Calculates fire behaviour prediction using CFFDRS. """
-    # pylint: disable=too-many-locals
-
-    # set default values in case the calculation fails (likely due to missing data)
-    fmc = cffdrs.foliar_moisture_content(latitude, longitude, elevation, get_julian_date_now())
-    sfc = cffdrs.surface_fuel_consumption(fuel_type, bui, ffmc, pc)
-    if cc is None:
-        cc = FUEL_TYPE_DEFAULTS[fuel_type]["CC"]
-
-    ros = cffdrs.rate_of_spread(FuelTypeEnum[fuel_type], isi, bui, fmc, sfc, pc=pc,
-                                cc=cc,
-                                pdf=pdf,
-                                cbh=cbh)
-    if sfc is not None:
-        cfb = calculate_cfb(FuelTypeEnum[fuel_type], fmc, sfc, ros, cbh)
-
-    if ros is not None and cfb is not None and cfl is not None:
-        hfi = cffdrs.head_fire_intensity(fuel_type=FuelTypeEnum[fuel_type],
-                                         percentage_conifer=pc,
-                                         percentage_dead_balsam_fir=pdf,
-                                         ros=ros, cfb=cfb, cfl=cfl, sfc=sfc)
-
-    lb_ratio = cffdrs.length_to_breadth_ratio(FuelTypeEnum[fuel_type], wind_speed)
-    wsv = cffdrs.calculate_wind_speed(FuelTypeEnum[fuel_type],
-                                      ffmc=ffmc,
-                                      bui=bui,
-                                      ws=wind_speed,
-                                      fmc=fmc,
-                                      sfc=sfc,
-                                      pc=pc,
-                                      cc=cc,
-                                      pdf=pdf,
-                                      cbh=cbh,
-                                      isi=isi)
-
-    bros = cffdrs.back_rate_of_spread(FuelTypeEnum[fuel_type],
-                                      ffmc=ffmc,
-                                      bui=bui,
-                                      wsv=wsv,
-                                      fmc=fmc, sfc=sfc,
-                                      pc=pc,
-                                      cc=cc,
-                                      pdf=pdf,
-                                      cbh=cbh)
-
-    sixty_minute_fire_size = get_fire_size(FuelTypeEnum[fuel_type], ros, bros, 60, cfb, lb_ratio)
-
-    fire_type = get_fire_type(FuelTypeEnum[fuel_type], crown_fraction_burned=cfb)
-
-    if hfi is not None:
-        intensity_group = calculate_intensity_group(hfi)
-
-    fire_behaviour_prediction = FireBehaviourPrediction(ros=ros, hfi=hfi, intensity_group=intensity_group,
-                                                        sixty_minute_fire_size=sixty_minute_fire_size,
-                                                        fire_type=fire_type)
-    return fire_behaviour_prediction
-
-
-def calculate_fire_behaviour_prediction_using_c7b(latitude: float,
-                                                  longitude: float,
-                                                  elevation: float,
-                                                  ffmc: float,
-                                                  bui: float,
-                                                  wind_speed: float,
-                                                  cc: float,  # pylint: disable=invalid-name
-                                                  cbh: float,
-                                                  cfl: float):
-    """ Calculates fire behaviour prediction using C7B. """
-
-    ros = c7b.rate_of_spread(ffmc=ffmc, bui=bui, wind_speed=wind_speed, percentage_slope=0.0, cc=cc)
-
-    fmc = cffdrs.foliar_moisture_content(latitude, longitude, elevation, get_julian_date_now())
-
-    sfc = cffdrs.surface_fuel_consumption(fuel_type=FuelTypeEnum.C7, bui=bui, ffmc=ffmc, pc=None)
-    cfb = cffdrs.crown_fraction_burned(fuel_type=FuelTypeEnum.C7, fmc=fmc,
-                                       sfc=sfc, ros=ros, cbh=cbh)
-
-    hfi = cffdrs.head_fire_intensity(fuel_type=FuelTypeEnum.C7,
-                                     percentage_conifer=None,
-                                     percentage_dead_balsam_fir=None,
-                                     ros=ros, cfb=cfb, cfl=cfl, sfc=sfc)
-
-    fire_type = get_fire_type(FuelTypeEnum.C7B, cfb)
-
-    intensity_group = calculate_intensity_group(hfi)
-
-    # TODO: not required for HFI, but for FireBat - we need to calculate 60 minute fire size, which
-    # will take a fair amount of peeking at the math.
-    # Some of the math in the c7b.rate_of_spread can be extracted, and the standard cffdrs math used.
-    fire_behaviour_prediction = FireBehaviourPrediction(
-        ros=ros,
-        hfi=hfi,
-        intensity_group=intensity_group,
-        sixty_minute_fire_size=None,
-        fire_type=fire_type)
-
-    return fire_behaviour_prediction
-
-
-def calculate_fire_behaviour_prediction(latitude: float,  # pylint: disable=too-many-arguments
-                                        longitude: float, elevation: float,
-                                        fuel_type: FuelTypeEnum,
-                                        bui: float, ffmc: float, wind_speed: float,
-                                        cc: float,  # pylint: disable=invalid-name
-                                        pc: float,  # pylint: disable=invalid-name
-                                        isi: float, pdf: float, cbh: float, cfl: float):
-    """ Calculate the fire behaviour prediction. """
-    if fuel_type == FuelTypeEnum.C7B:
-        return calculate_fire_behaviour_prediction_using_c7b(
-            latitude=latitude,
-            longitude=longitude,
-            elevation=elevation,
-            ffmc=ffmc,
-            bui=bui,
-            wind_speed=wind_speed,
-            cc=cc,
-            cbh=cbh,
-            cfl=cfl)
-    return calculate_fire_behaviour_prediction_using_cffdrs(
-        latitude=latitude,
-        longitude=longitude,
-        elevation=elevation,
-        fuel_type=fuel_type,
-        bui=bui,
-        ffmc=ffmc,
-        cc=cc,
-        pc=pc,
-        wind_speed=wind_speed,
-        isi=isi,
-        pdf=pdf,
-        cbh=cbh,
-        cfl=cfl)
-
-
-# TODO: this doesn't belong in wildfire_one - it's doing fire behaviour calculations
-def generate_station_daily(raw_daily,  # pylint: disable=too-many-locals
-                           station: WFWXWeatherStation,
-                           fuel_type: FuelType) -> StationDaily:
-    """ Transform from the raw daily json object returned by wf1, to our daily object.
-    """
-    # pylint: disable=invalid-name
-    # we use the fuel type lookup to get default values.
-    pc = fuel_type.percentage_conifer if fuel_type.percentage_conifer is not None\
-        else FUEL_TYPE_DEFAULTS[fuel_type.fuel_type_code]["PC"]
-    pdf = fuel_type.percentage_dead_fir if fuel_type.percentage_dead_fir is not None\
-        else FUEL_TYPE_DEFAULTS[fuel_type.fuel_type_code]["PDF"]
-    cbh = FUEL_TYPE_DEFAULTS[fuel_type.fuel_type_code]["CBH"]
-    cfl = FUEL_TYPE_DEFAULTS[fuel_type.fuel_type_code]["CFL"]
-    date = raw_daily.get('weatherTimestamp', None)
-    isi = raw_daily.get('initialSpreadIndex', None)
-    bui = raw_daily.get('buildUpIndex', None)
-    ffmc = raw_daily.get('fineFuelMoistureCode', None)
-    cc = raw_daily.get('grasslandCuring', None)
-    wind_speed = raw_daily.get('windSpeed', None)
-
-    try:
-        fire_behaviour_prediction = calculate_fire_behaviour_prediction(
-            latitude=station.lat,
-            longitude=station.long,
-            elevation=station.elevation,
-            fuel_type=FuelTypeEnum[fuel_type.fuel_type_code],
-            bui=bui,
-            ffmc=ffmc,
-            wind_speed=wind_speed,
-            cc=cc,
-            pc=pc,
-            isi=isi,
-            pdf=pdf,
-            cbh=cbh,
-            cfl=cfl)
-    # pylint: disable=broad-except
-    except Exception as exc:
-        logger.error('Encountered error while generating StationDaily for station %s', station.code)
-        logger.error(exc, exc_info=True)
-        # prediction calculation failed, so we set the values to None
-        fire_behaviour_prediction = FireBehaviourPrediction(None, None, None, None, None)
-
-    return StationDaily(
-        code=station.code,
-        date=date,
-        status=raw_daily.get('recordType', '').get('id', None),
-        temperature=raw_daily.get('temperature', None),
-        relative_humidity=raw_daily.get('relativeHumidity', None),
-        wind_speed=raw_daily.get('windSpeed', None),
-        wind_direction=raw_daily.get('windDirection', None),
-        precipitation=raw_daily.get('precipitation', None),
-        grass_cure_percentage=raw_daily.get('grasslandCuring', None),
-        ffmc=ffmc,
-        dmc=raw_daily.get('duffMoistureCode', None),
-        dc=raw_daily.get('droughtCode', None),
-        fwi=raw_daily.get('fireWeatherIndex', None),
-        # Danger class possible values: 1, 2, 3, 4, and 5. Where 1 is the lowest, and 5 is the highest
-        # This is the same for the dangerGrassland and dangerScrub,
-        # but those values aren’t really used anywhere,
-        # just calculated and stored along with the forest danger rating
-        # You can see current stations/rating on this page here:
-        # https://wfapps.nrs.gov.bc.ca/pub/wfwx-danger-summary-war/dangerSummary
-        danger_class=raw_daily.get('dangerForest', None),
-        isi=isi,
-        bui=bui,
-        rate_of_spread=fire_behaviour_prediction.ros,
-        hfi=fire_behaviour_prediction.hfi,
-        observation_valid=raw_daily.get('observationValidInd', None),
-        observation_valid_comment=raw_daily.get(
-            'observationValidComment', None),
-        intensity_group=fire_behaviour_prediction.intensity_group,
-        sixty_minute_fire_size=fire_behaviour_prediction.sixty_minute_fire_size,
-        fire_type=fire_behaviour_prediction.fire_type,
-        error=raw_daily.get('observationValidInd', None),
-        error_message=raw_daily.get('observationValidComment', None),
-        last_updated=datetime.fromtimestamp(raw_daily.get(
-            'lastEntityUpdateTimestamp') / 1000, tz=timezone.utc)
     )
